pyproject.toml
[tool.poetry]
name = "chatbot-rag"
version = "0.1.0"
description = ""
authors = ["eng-alecrim <lucasalecrimvieira@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.1"
langchain = "^0.2.10"
langchain-openai = "^0.1.17"
tiktoken = "^0.7.0"
streamlit = "^1.36.0"
langchain-community = "^0.2.10"
faiss-cpu = "^1.8.0.post1"
pypdf = "^4.3.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.1"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"
ruff = "^0.5.4"
ignr = "^2.2"
black = "^24.4.2"
jupyterlab = "^4.2.4"
catppuccin-jupyterlab = "^0.2.3"

[tool.ruff]
line-length = 79
extend-exclude = ["migrations"]

[tool.ruff.lint]
preview = true
select = ["I", "F", "E", "W", "PL", "PT"]
ignore = ["E501", "E731", "PLR1711"]

[tool.ruff.format]
indent-style = "space"
preview = true
quote-style = "double"

[tool.pytest.ini_options]
pythonpath = "."
addopts = "-p no:warnings"

[tool.taskipy.tasks]
lint = "ruff check . && ruff check . --diff"
format = "ruff check . --fix && ruff format ."
pre_test = "task lint"
test = "pytest -s -x --cov=chatbot_rag -vv"
post_test = "coverage html"
jupyter = "python -m jupyterlab"
run = "streamlit run chatbot_rag/app.py"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
